Add files via upload
The Tip Calculator project is designed to provide users with an easy way to calculate tips based on the amount of their bill. The user will enter the total amount of their bill, select the percent they would like to tip, and click the “Calculate” button. The program will then display the tip amount and the total bill amount which includes the tip amount. The Tip Calculator project also includes a feature to help user to divide their tips 
Equally to all .